feat: extract turboapi-core shared Zig library (router + HTTP utils)
Extract the radix trie router and HTTP utility functions from zig/src/
into a standalone turboapi-core/ library so both turboAPI (Python) and
merjs (Zig full-stack) can share the same routing and HTTP primitives.

- turboapi-core/: new zero-dep Zig library with router, http utils,
  bounded cache, and shared types (23 tests, fuzz-tested)
- zig/src/router.zig removed — now imported from turboapi-core
- zig/src/server.zig, main.zig, db.zig updated to use turboapi-core
- All 248 Python tests pass, 134k req/s benchmark unchanged
