+++ turboapi-core/.zig-cache/o/f1ea4330472a4c6ebcbfed6922919020/dependencies.zig
@@ -1,0 +1,2 @@
+pub const packages = struct {};
+pub const root_deps: []const struct { []const u8, []const u8 } = &.{};
