+++ turboapi-core/build.zig.zon
@@ -1,0 +1,12 @@
+.{
+    .name = .turboapi_core,
+    .fingerprint = 0x160efbab8247370e,
+    .version = "0.1.0",
+    .minimum_zig_version = "0.15.0",
+    .dependencies = .{},
+    .paths = .{
+        "build.zig",
+        "build.zig.zon",
+        "src",
+    },
+}
